requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==3.0.1
constantly==15.1.0
cryptography==39.0.1
daphne==4.0.0
Django==4.1.5
django-admin==2.0.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-excel-base==1.0.5
django-excel-response2==3.0.3
django-six==1.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.4
phonenumbers==8.13.4
Pillow==9.4.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
python-dateutil==2.8.2
pytz==2022.7.1
razorpay==1.3.0
redis==4.5.1
requests==2.28.2
screen==1.0.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
twilio==7.16.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
urllib3==1.26.14
xlwt==1.3.0
zope.interface==5.5.2